news&sentiment_Analyst/requirements.txt
feedparser
requests
praw
transformers
torch
python-dotenv

